feat: plugin css & clean up build scripts
