+++ package.json
@@ -13,7 +13,8 @@
     "dependencies": {
         "console-menu": "^0.1.0",
         "discord-types": "^1.3.26",
-        "electron-devtools-installer": "^3.2.0"
+        "electron-devtools-installer": "^3.2.0",
+        "esbuild-sass-plugin": "^2.3.3"
     },
     "scripts": {
         "buildWeb": "node buildWeb.mjs",
