--- src/pluginsModule.d.ts
@@ -1,9 +1,0 @@
-declare module "plugins" {
-    const plugins: Record<string, import("./utils/types").Plugin>;
-    export default plugins;
-}
-
-declare module "git-hash" {
-    const hash: string;
-    export default hash;
-}
