+++ src/utils/types.ts
@@ -36,6 +36,7 @@ interface PluginDef {
     commands?: Command[];
     dependencies?: string[],
     required?: boolean;
+    css?: string;
     /**
      * Set this if your plugin only works on Browser or Desktop, not both
      */
